mrdegibbs: clean up incoming code
- add copyright notice, author info, requisite usage() fields
- use Math::pi
- some style changes
